FINALLY fixed scroll jumping!
Added lastTouchPt and use it to "cover" where the hardware says it has a touch but actually doesn't so what you get back is batty! If you check the z value it will flag a bad touch so, insert the last saved good touch point and you are good to go!

This happens at the end of  touch where you get a true that there is a touch but by the time you grab the point there is no touch. And it happened a LOT. Causing scrolling to jump to the end of things right as you had it set where you wanted.